requirements.txt
click==7.1.2
Flask==1.1.4
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
numpy==1.22.1
psycopg2==2.9.3
SQLAlchemy==1.3.23
urllib3==1.25.9
uWSGI==2.0.20
Werkzeug==1.0.1
WTForms==3.0.1
opencensus-ext-azure
sentry-sdk
pytest
selenium